Type of does-not-contain search depends of query
This will do a regex search, if query is 'anywhere' or 'name',
but it will do an exact not-match in 'tags' and 'projects'.
This is useful as names are often a concatenation of multiple
strings and exact matches are unlikely.
Tags and Projects on the other side are short defined strings,
and exact not-matches are more useful.